pyproject.toml
[project]
name = "inv-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.17.2",
    "black>=25.12.0",
    "celery>=5.6.0",
    "email-validator>=2.3.0",
    "faker>=38.2.0",
    "fastapi>=0.124.0",
    "fastapi-cors>=0.0.6",
    "flake8>=7.3.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-barcode>=0.16.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "python-socketio>=5.15.0",
    "qrcode>=8.2",
    "redis>=7.1.0",
    "reportlab>=4.4.5",
    "sentry-sdk>=2.47.0",
    "sqlalchemy>=2.0.44",
    "supabase>=2.25.1",
    "uvicorn[standard]>=0.38.0",
]
